requirements.txt
deepgram-sdk==3.7.0
speechmatics-flow
pymilvus>=2.2.0
numpy>=1.21.0
python-dotenv>=0.19.0
sounddevice>=0.4.5
soundfile>=0.10.3
pyaudio>=0.2.11
google-generativeai>=0.8.0
sentence-transformers>=2.2.0
websocket-client

# PDF Processing Dependencies
PyPDF2>=3.0.0
pdfplumber>=0.10.0
